Updated small fixes and 2 Interfaces
